+++ source/ada/lsp-ada_configurations.adb
@@ -432,10 +432,10 @@ package body LSP.Ada_Configurations is
             then
                Self.Range_Formatting_Fallback := JSON (Index).Boolean_Value;
 
-            elsif Name = "showNotificationsOnErrors"
+            elsif Name = "showNotificationsOnErrors" or else Name = "metricThresholds"
             then
-               --  This is a VS Code only setting, treated at the VS Code
-               --  extension's level. We still include it here to mark it as
+               --  These are VS Code only settings, treated at the VS Code
+               --  extension's level. We still include them here to mark them as
                --  recognized and to support the settings-doc test that checks
                --  that each setting is documented.
                null;